fix: Avoid division by 0 failure
The prior fix triggers `Asso_Leg()` `ZeroDivisionError` condition for inputs such as `L2M2`, perhaps this is why the range was being truncated before?

I've replaced the method for calculating Legendre Polynomials with a single method and slightly changed the if/else logic in the `Y` value calculation loops to support it. The method is ported from [this rust project](https://github.com/Jasper-Bekkers/spherical-harmonics/blob/master/src/lib.rs#L57) and appears to perform the calculations correctly.

I don't understand the math well enough for this calculation to update the existing python code instead. This does avoid the error and correctly graph the result(that otherwise lacks colour/fill).